# swagger.yaml
openapi: 3.0.0
info:
  title: Open Finance Project API
  description: API documentation for the Open Finance Project integrated with the Open Bank Project (OBP-API).
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local server
paths:
...
  /api/accounts/{accountNumber}/balance:
    get:
      summary: Get Account Balance (Custom Endpoint)
      description: Returns the balance for a given account in your Open Finance project.
      parameters:
        - name: accountNumber
          in: path
          required: true
          description: The account number.
          schema:
            type: string
      responses:
        '200':
          description: The balance of the account.
          content:
            application/json:
              schema:
                type: object
                properties:
                  balance:
                    type: number
                    example: 1000.0
  /api/accounts/{accountNumber}/deposit:
    post:
      summary: Deposit to Account
      description: Deposits a specified amount to an account.
      parameters:
        - name: accountNumber
          in: path
          required: true
          description: The account number.
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                amount:
                  type: number
                  description: Amount to deposit.
                  example: 500
      responses:
        '200':
          description: The updated account after the deposit.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 1
                  accountNumber:
                    type: string
                    example: "123456"
                  balance:
                    type: number
                    example: 1500.0
...